track connected peers in a hashset

+++ fuel-p2p/src/config.rs
@@ -28,7 +28,7 @@ pub struct P2PConfig {
     // `DiscoveryBehaviour` related fields
     pub bootstrap_nodes: Vec<(PeerId, Multiaddr)>,
     pub enable_mdns: bool,
-    pub max_peers_connected: u64,
+    pub max_peers_connected: usize,
     pub allow_private_addresses: bool,
     pub enable_random_walk: bool,
     pub connection_idle_timeout: Option<Duration>,
